verbs: Always allocate a verbs_context
Now that everything is in one tree we can revise the legacy init_context
path to always allocate a verbs_context by swapping the ibv_context
for a verbs_context in all of the provider's wrapper struct.

To keep the provider diffs minimal this single patch does several things
at once:

- Introduce the verbs_init_and_alloc_context() macro.
  This allocates, zeros and initializes the verbs_context for each driver.
  Notably this new macro correctly sets errno as required upon failure.

- Remove boilerplate from all drivers, calloc, malloc, memset, cmd_fd and
  device assignment

- Along with the verbs_init scheme necessarily comes the verbs_uninit
  scheme which lowers the uninit call into the provder not the common
  code. This allows us to properly uninit on the init error path.

Together this follows the fairly successful pattern we see in the kernel
for driver init to a subsystem.

Also this changes ibv_cmd_get_context to accept a verbs_context since
the majority of callers are now providing that, this keeps the diff
smaller.

This makes the entire flow more consistent and will let us eliminate the
init_context flow.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>