verbs: Add basic infrastructure for mixed write and ioctl cmds
This provides a series of helper macros designed to let us merge the ioctl and
write paths of each command into a single function following a standard
pattern.

The macros intend to support a universal cmd function that accepts a
ibv_commad_buffer link from the driver, allowing the driver to provide
both the classic UHW structures as well as any additional driver specific
ids required.

To avoid changing every driver a wrapper is provided that generates the
ibv_command_buffer link from the standard cmd and resp pointers.

The helper macros are designed so the ioctl work can be largely elided by the
compiler if ioctl support is disabled (not done yet).

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>